Major improvements (scrape function + scrape_vg)
Fully functional scrape function for metacritic website. Mostly functioning function for scraping vgchartz. Current drawback is 429 error run into. Current script also creates meta_urls, which is a list of the urls on the metacritic website that are relevant to the project. metacritic.csv is the scraped data in tabular form.